+++ ReleaseNotes.md
@@ -10,11 +10,12 @@ The control panel is installed once. It will automatically download the correct 
 
 - 3.3.0: Improved process kill
   - will now attempt to locate and kill a process using the port even if the PID file is stale
+  - SIGINT, SIGTERM, SIGKILL are treated as intentional stops, same as using the stop button, no restarts.
 
 - 3.3.0: Command-line options for multiple instances and Linux daemon mode
   - Run controlpanel --help for details
   - These options are Linux-oriented, targeted at virtual privatehosting scenarios.
-  - running with --owlcms --tracker both creates a connected tandem where owlcms feeds the tracker on the port indicated by the tracker config.Add
+  - running with --owlcms --tracker both creates a connected tandem where owlcms feeds the tracker on the port indicated by the tracker config.
   - controlpanel --init now initializes the main owlcms instance instead of reporting an empty instance name
   - A daemon mode is provided
     - Under systemd, the Go process stays alive and supervises OWLCMS (restart on non-zero exit).